Updates to the MPIService
Various updates to the MPIService:
  - split MPIService into header, implementation and plugin;
  - add comments about the implementation status of MPI_THREAD_MULTIPLE;
  - add helper function to make sure the MPIService is available;
  - add an option to set the pmix_server_uri OpenMPI parameter.
