+++ HeterogeneousCore/MPIServices/BuildFile.xml
@@ -1,0 +1,6 @@
+<use name="openmpi"/>
+<use name="FWCore/Framework"/>
+<use name="FWCore/ParameterSet"/>
+<export>
+  <lib name="1"/>
+</export>
